Using threads to keep UI responsive while flashing
Since StreamReader.Peek() will hang on empty buffer, Application.DoEvents(); will never be reached until the flashing process is finished.
This will cause the UI to be unresponsive during firmware flashing, especially for flashing utilities that take more time to finish. (such as bootloadHID)

This approach starts a thread for firmware flashing, which creates two additional threads for processing messages in STDOUT and STDERR produced by the flashing utility.

There may be a better way to do this, though.