[Windows] Fix crash when a HID device has no input reports
Some devices, like the XD002 keyboard with the factory firmware, have
HID report descriptors which do not declare any input reports.  If such
HID descriptor additionally had the vendor-defined page and usage codes
which correspond to the HID console, QMK Toolbox tried to use that HID
interface as a HID console, and an attempt to read the nonexistent input
report caused a crash inside HidLibrary; this made using QMK Toolbox
impossible when a device with such HID report descriptor is connected.

Add a check for a non-zero input report size to GetListableDevices() to
prevent the attempt to use a HID interface without any input repors as a
HID console.