Recognize 2A03:0040 (Arduino Leonardo ETH) as a Caterina bootloader
Apparently some people sell Pro Micro clones with a variant of the
Caterina bootloader which shows up as 2A03:0040; that VID:PID was
originally used by the bootloader on the Arduino Leonardo ETH board.
Add code to recognize that VID:PID combination as Caterina, so that
those Pro Micro clones could be flashed.